Move shutdown signal installer to control flow
